Segregação do componente ChoosePostSteps
