+++ src/main/components/draw/pages/FacebookDraw.jsx
@@ -28,12 +28,12 @@ const FacebookDraw = (props) => {
             <div className="container">
                 <div className="row">
                     <div className="col-12 col-md-10 offset-md-1">
-                        <h1 className="h2">Sorteio de Comentário do Facebook</h1>
-                        <h2 className="h3">Siga os passos a seguir para fazer o sorteio</h2>
+                        <h1 className="h2">Sorteio de Comentários do Facebook</h1>
                     </div>
                 </div>
                 <div className="row mt-5">
                     <div className="col-12 col-md-10 offset-md-1">
+                        <h2 className="h4 mb-4">Siga os passos a seguir para fazer o sorteio</h2>
                         <ChoosePostSteps />
                     </div>
                 </div>
